[XY]update - button style, page height and Q4 file name
