+++ src/components/Questions.css
@@ -1,12 +1,9 @@
-body {
-  margin: 0;
-}
-
 .border-box-qa {
-  display: flex;
-  flex-wrap: wrap;
   height: 100vh;
-  overflow: hidden;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
 }
 
 .qa-label,
@@ -78,16 +75,43 @@ body {
   flex-direction: column;
 }
 
+hr {
+  border-bottom: 1px solid #fff;
+  width: 40%;
+  margin: 40px 0;
+}
+
+/* Basic button styles */
 .button-qa {
   padding: 10px 20px;
   margin: 10px 3px;
-  background-color: #0018A4;
   color: #fff;
+  text-align: center;
+  font-size: 1rem;
   font-weight: 600;
   border-radius: 6px;
-  border: none;
+  border: solid 1px black;
   cursor: pointer;
+  display: inline-block;
+  width: auto;
+  transition: background-color 0.3 ease, color 0.3s ease;
 }
+
+/* Default button style */
+.button-qa {
+  background-color: #c2c2c2;
+}
+
+/* Active button style */
+.button-qa:active {
+  background-color: #0018A4;
+}
+
+/* Hover state for better UX */
+.button-qa:hover {
+  opacity: 0.85;
+}
+
 .cta-box {
   position: absolute;
   bottom: 20%;
