+++ src/components/Welcome.css
@@ -1,31 +1,31 @@
 .header-welcome-form {
-    height: 100vh;
-    /* overflow: hidden; */
-    background-color: #0018A4;
-    color: white;
-    text-align: center;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    padding: 0 30px;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  background-color: #0018A4;
+  color: white;
+  text-align: center;
 }
+
 .center-content-header {
-    height: 50%;
-    width: 50%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
+  height: 50%;
+  width: 50%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
 }
+
 .button-header {
-    padding: 10px 20px;
-    margin: 20px 0;
-    color: #0018A4;
-    font-weight: 600;
-    border-radius: 6px;
-    background-color: #fff;
-    border: none;
-    cursor: pointer;
+  padding: 10px 20px;
+  margin: 20px 0;
+  color: #0018A4;
+  font-weight: 600;
+  border-radius: 6px;
+  background-color: #fff;
+  border: solid 1px black;
+  cursor: pointer;
 }
