+++ src/index.css
@@ -8,8 +8,11 @@
   -moz-osx-font-smoothing: grayscale;
 }
 
-body {
+body, html {
   margin: 0;
+  padding: 0;
+  height: 100%;
+  overflow: hidden;
 }
 
 code {
